fix: satisfy workspace handoff static and rust checks
